Added health bar and manager
