+++ Assets/Quests/EatYellowFlower/EatYellowFlower.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8074391951f0fd64faf51262d774e911, type: 3}
   m_Name: EatYellowFlower
   m_EditorClassIdentifier: 
-  _name: TestQuest
+  _name: Yellow Flower
   _description: This is a test quest
   _startAmount: 0
   _endAmount: 1
@@ -24,5 +24,5 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 096044419b70fd542a302b1555045490, type: 2}
   _startQuestsOnCompletion:
   - {fileID: 0}
-  _currentAmount: 1
-  _currentState: 4
+  _currentAmount: 0
+  _currentState: 1
