+++ Assets/Scripts/Interactables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 562ce93243e72c54eb8faf5ab94c37b0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
