+++ Assets/Scripts/Interactables/Consumable.cs
