+++ Assets/Scripts/Interactables/Consumable.cs.meta
