+++ Assets/Scripts/Interactables/IInteractableObj.cs.meta
