+++ Assets/Scripts/Interactables/RedFlowerController.cs
